+++ src/client/components/Input/Input-styled.tsx
@@ -9,7 +9,6 @@ const StyledInputContainer = styled.div<StyledInputContainerProps>`
   flex-direction: column;
   margin: ${(props) => (props.full ? "0" : undefined)};
   margin-bottom: 25px;
-  margin-top: 32px;
   max-width: ${(props) => (props.full ? `100%` : `90%`)};
   width: ${(props) => (props.full ? `100%` : ``)};
 
@@ -23,7 +22,7 @@ const StyledInputInnerContainer = styled.div`
   box-shadow: 0px 2px 4px rgba(0, 0, 0, 0.05);
   display: flex;
   flex-wrap: wrap;
-  margin-top: 14px;
+  margin-top: 5px;
   width: 100%;
 
   input {
@@ -79,7 +78,6 @@ const StyledInputLabel = styled.label`
   font-size: 14px;
   font-weight: bold;
   line-height: 21px;
-  text-transform: uppercase;
 `;
 
 export {