+++ src/client/components/Profile/Profile-styled.tsx
@@ -17,12 +17,6 @@ const ProfileAccountDetailsContainer = styled.div`
 const ProfileAccountDetailsContentContainer = styled.div`
   padding: 25px;
 
-  h4 {
-    font-size: 14px;
-    margin: 0;
-    text-transform: uppercase;
-  }
-
   input {
     :disabled {
       color: #b9bdcf;
@@ -57,12 +51,8 @@ const ProfileAccountDetailsHeadingContainer = styled.div`
   }
 `;
 
-const ProfileAccountStatsAvatar = styled.img`
-  background-image: linear-gradient(135deg, #ce9ffc 10%, #7367f0 100%);
-  border-radius: 50%;
-  height: 150px;
+const ProfileAccountStatsAvatarContainer = styled.div`
   margin: -75px auto;
-  width: 150px;
 `;
 
 const ProfileAccountStatsContainer = styled.div`
@@ -146,7 +136,7 @@ export {
   ProfileAccountDetailsContentContainer,
   ProfileAccountDetailsHeadingContainer,
   ProfileAccountStatsActionsContainer,
-  ProfileAccountStatsAvatar,
+  ProfileAccountStatsAvatarContainer,
   ProfileAccountStatsContainer,
   ProfileAccountStatsInnerContainer,
   ProfileHeroContainer,